(#1) Introduce file compression
