+++ src/main/kotlin/Main.kt
@@ -1,29 +1,11 @@
-import java.io.IOException
-import java.nio.file.*
-import java.nio.file.attribute.BasicFileAttributes
+import java.nio.file.Files
+import java.nio.file.Paths
 
 fun main(args: Array<String>) {
-    val directory = Paths.get(args[0])
-    Files.walkFileTree(directory, object : FileVisitor<Path> {
-        override fun visitFile(file: Path?, attrs: BasicFileAttributes?): FileVisitResult {
-            println("Processing ${file.toString()}")
-            return FileVisitResult.CONTINUE
-        }
+    val outputDir = Paths.get("/home/krzysztof/Videos")
 
-        override fun preVisitDirectory(dir: Path?, attrs: BasicFileAttributes?): FileVisitResult {
-            println("Visits ${dir.toString()}")
-            return FileVisitResult.CONTINUE
-        }
+    val rootDirectory = Paths.get(args[0])
 
-        override fun visitFileFailed(file: Path?, exc: IOException?): FileVisitResult {
-            TODO("Not yet implemented")
-        }
-
-        override fun postVisitDirectory(dir: Path?, exc: IOException?): FileVisitResult {
-            println("Finished ${dir.toString()}")
-            return FileVisitResult.CONTINUE
-        }
-
-    })
+    Files.walkFileTree(rootDirectory, VideoVisitor(rootDirectory, outputDir))
 
 }
